+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Version 0.0.9
 
 + Support migrate from WordPress
 + Support customize post permalink
++ Support customize tag permalink
 
 Version 0.0.8
 --------------